Final Refactor, moving robot main functions to subclasses. SJD
